+++ tests/psimulate/test_cli.py
@@ -1,14 +1,7 @@
-"""Tests for the --run-config YAML configuration file feature.
-
-These tests verify that:
-- A YAML config file can supply CLI argument values via --run-config
-- CLI arguments override values from the config file
-- Validation (Click types, callbacks, Choice) applies to config file values
-- Unknown keys in the config file produce clear errors
-- Malformed YAML produces a clear error
-- The feature works for the run, restart, and expand subcommands
+"""Tests for psimulate CLI features.
 """
 
+import datetime
 from pathlib import Path
 from typing import Any
 from unittest.mock import patch
@@ -73,11 +66,6 @@ def result_dir(tmp_path: Path) -> Path:
 def results_root(tmp_path: Path) -> Path:
     """A writable directory that can serve as results_root."""
     return _make_results_root(tmp_path)
-
-
-# ===================================================================
-# Tests for ``psimulate run --run-config``
-# ===================================================================
 
 
 class TestRunSubcommand:
@@ -214,11 +202,6 @@ class TestRunSubcommand:
         mock_main.assert_called_once()
 
 
-# ===================================================================
-# Tests for ``psimulate restart --run-config``
-# ===================================================================
-
-
 class TestRestartSubcommand:
     """Tests for --run-config on the ``restart`` subcommand."""
 
@@ -268,11 +251,6 @@ class TestRestartSubcommand:
         mock_main.assert_called_once()
         call_kwargs = mock_main.call_args.kwargs
         assert call_kwargs["native_specification"].peak_memory == 8
-
-
-# ===================================================================
-# Tests for ``psimulate expand --run-config``
-# ===================================================================
 
 
 class TestExpandSubcommand:
@@ -333,11 +311,6 @@ class TestExpandSubcommand:
         assert call_kwargs["extra_args"]["num_seeds"] == 3  # from config
 
 
-# ===================================================================
-# Error handling tests
-# ===================================================================
-
-
 class TestErrorHandling:
     """Tests for error cases in --run-config."""
 
@@ -368,11 +341,6 @@ class TestErrorHandling:
         result = cli_runner.invoke(psimulate, ["run", "--run-config", str(config)])
         assert result.exit_code != 0
         assert "bogus_option" in result.output
-
-
-# ===================================================================
-# Validation parity tests — config values go through same validation
-# ===================================================================
 
 
 class TestValidationParity:
@@ -416,3 +384,235 @@ class TestValidationParity:
         result = cli_runner.invoke(psimulate, ["run", "--run-config", str(config)])
         assert result.exit_code != 0
         assert "nonexistent_node_type" in result.output or "not supported" in result.output
+
+
+# The date after which positional argument support should be removed.
+_DEPRECATION_DEADLINE = datetime.date(2026, 10, 9)
+
+
+class TestKeywordArgs:
+    """Using the new keyword options works without deprecation warnings."""
+
+    def test_run_keyword_args(
+        self, tmp_path: Path, model_spec: Path, branch_config: Path, result_dir: Path
+    ) -> None:
+        """model_specification and branch_configuration via --flags."""
+        cli_runner = CliRunner()
+        with patch(_RUNNER_MAIN) as mock_main:
+            result = cli_runner.invoke(
+                psimulate,
+                [
+                    "run",
+                    "--model-specification",
+                    str(model_spec),
+                    "--branch-configuration",
+                    str(branch_config),
+                    "--result-directory",
+                    str(result_dir),
+                    "--project",
+                    "proj_simscience",
+                ],
+            )
+
+        assert result.exit_code == 0, result.output
+        assert "deprecated" not in result.output.lower()
+        mock_main.assert_called_once()
+        call_kwargs = mock_main.call_args.kwargs
+        assert call_kwargs["input_paths"].model_specification == model_spec.resolve()
+        assert call_kwargs["input_paths"].branch_configuration == branch_config.resolve()
+
+    def test_run_short_flags(
+        self, tmp_path: Path, model_spec: Path, branch_config: Path, result_dir: Path
+    ) -> None:
+        """Short flags -f and -b work for model spec and branches."""
+        cli_runner = CliRunner()
+        with patch(_RUNNER_MAIN) as mock_main:
+            result = cli_runner.invoke(
+                psimulate,
+                [
+                    "run",
+                    "-M",
+                    str(model_spec),
+                    "-B",
+                    str(branch_config),
+                    "-o",
+                    str(result_dir),
+                    "-P",
+                    "proj_simscience",
+                ],
+            )
+
+        assert result.exit_code == 0, result.output
+        assert "deprecated" not in result.output.lower()
+        mock_main.assert_called_once()
+
+    def test_restart_keyword_arg(self, tmp_path: Path, results_root: Path) -> None:
+        """results_root via --results-root flag."""
+        cli_runner = CliRunner()
+        with patch(_RUNNER_MAIN) as mock_main:
+            result = cli_runner.invoke(
+                psimulate,
+                [
+                    "restart",
+                    "--results-root",
+                    str(results_root),
+                    "--project",
+                    "proj_simscience",
+                ],
+            )
+
+        assert result.exit_code == 0, result.output
+        assert "deprecated" not in result.output.lower()
+        mock_main.assert_called_once()
+        call_kwargs = mock_main.call_args.kwargs
+        assert call_kwargs["input_paths"].result_directory == results_root.resolve()
+
+    def test_expand_keyword_arg(self, tmp_path: Path, results_root: Path) -> None:
+        """results_root via --results-root flag on expand."""
+        cli_runner = CliRunner()
+        with patch(_RUNNER_MAIN) as mock_main:
+            result = cli_runner.invoke(
+                psimulate,
+                [
+                    "expand",
+                    "--results-root",
+                    str(results_root),
+                    "--project",
+                    "proj_simscience",
+                ],
+            )
+
+        assert result.exit_code == 0, result.output
+        assert "deprecated" not in result.output.lower()
+        mock_main.assert_called_once()
+
+
+class TestPositionalArgDeprecationWarning:
+    """Using the old positional arguments still works but emits a warning."""
+
+    def test_run_positional_args_emit_warning(
+        self, tmp_path: Path, model_spec: Path, branch_config: Path, result_dir: Path
+    ) -> None:
+        cli_runner = CliRunner()
+        with pytest.warns(FutureWarning, match="deprecated"):
+            with patch(_RUNNER_MAIN) as mock_main:
+                result = cli_runner.invoke(
+                    psimulate,
+                    [
+                        "run",
+                        str(model_spec),
+                        str(branch_config),
+                        "--result-directory",
+                        str(result_dir),
+                        "--project",
+                        "proj_simscience",
+                    ],
+                )
+
+        assert result.exit_code == 0, result.output
+        mock_main.assert_called_once()
+        call_kwargs = mock_main.call_args.kwargs
+        assert call_kwargs["input_paths"].model_specification == model_spec.resolve()
+        assert call_kwargs["input_paths"].branch_configuration == branch_config.resolve()
+
+    def test_restart_positional_arg_emits_warning(
+        self, tmp_path: Path, results_root: Path
+    ) -> None:
+        cli_runner = CliRunner()
+        with pytest.warns(FutureWarning, match="deprecated"):
+            with patch(_RUNNER_MAIN) as mock_main:
+                result = cli_runner.invoke(
+                    psimulate,
+                    [
+                        "restart",
+                        str(results_root),
+                        "--project",
+                        "proj_simscience",
+                    ],
+                )
+
+        assert result.exit_code == 0, result.output
+        mock_main.assert_called_once()
+
+    def test_expand_positional_arg_emits_warning(
+        self, tmp_path: Path, results_root: Path
+    ) -> None:
+        cli_runner = CliRunner()
+        with pytest.warns(FutureWarning, match="deprecated"):
+            with patch(_RUNNER_MAIN) as mock_main:
+                result = cli_runner.invoke(
+                    psimulate,
+                    [
+                        "expand",
+                        str(results_root),
+                        "--project",
+                        "proj_simscience",
+                    ],
+                )
+
+        assert result.exit_code == 0, result.output
+        mock_main.assert_called_once()
+
+
+class TestPositionalKeywordConflict:
+    """Providing the same file path as both a positional arg and a keyword option
+    raises an error."""
+
+    def test_run_model_spec_conflict(
+        self, tmp_path: Path, model_spec: Path, branch_config: Path, result_dir: Path
+    ) -> None:
+        cli_runner = CliRunner()
+        result = cli_runner.invoke(
+            psimulate,
+            [
+                "run",
+                str(model_spec),  # positional
+                "--model-specification",
+                str(model_spec),  # keyword
+                "--branch-configuration",
+                str(branch_config),
+                "--result-directory",
+                str(result_dir),
+                "--project",
+                "proj_simscience",
+            ],
+        )
+        assert result.exit_code != 0
+        assert (
+            "model_specification" in result.output.lower()
+            or "conflict" in result.output.lower()
+        )
+
+    def test_restart_results_root_conflict(self, tmp_path: Path, results_root: Path) -> None:
+        cli_runner = CliRunner()
+        result = cli_runner.invoke(
+            psimulate,
+            [
+                "restart",
+                str(results_root),  # positional
+                "--results-root",
+                str(results_root),  # keyword
+                "--project",
+                "proj_simscience",
+            ],
+        )
+        assert result.exit_code != 0
+
+
+class TestDeprecationDeadline:
+    """This test fails after the deprecation deadline to remind us to remove
+    positional argument support entirely."""
+
+    def test_positional_args_deadline_not_reached(self) -> None:
+        """If this test fails, it is time to remove positional argument support
+        for model_specification, branch_configuration, and results_root.
+        Replace them entirely with the keyword options --model-specification/-M,
+        --branch-configuration/-B, and --results-root/-R."""
+        today = datetime.date.today()
+        assert today < _DEPRECATION_DEADLINE, (
+            f"Positional argument deprecation deadline ({_DEPRECATION_DEADLINE}) has passed. "
+            f"Remove positional argument support for model_specification, "
+            f"branch_configuration, and results_root. Use the keyword options "
+            f"--model-specification/-M, --branch-configuration/-B, and "
+            f"--results-root/-R exclusively."
+        )
